test_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file test_requirements.txt test_requirements.in
#
attrs==17.4.0             # via pytest
flake8==3.5.0
isort==4.3.4
mccabe==0.6.1             # via flake8
pluggy==0.6.0             # via pytest
py==1.5.2                 # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pytest-django==3.1.2
pytest==3.4.2
six==1.11.0               # via pytest
